✨ Start implementation for more flexible Compliance Ready Policies.
This commit offers a simple way to inject a policy configuratrion schema into
a policy.

Based on the policy pack configuration, Policies leveraging `.shouldEvalPolicy()`
will be able to determine if the given resource should skip the policy evaluation
or not. For now, the Pulumi resource name is the element used to determine whether
the evaluation shouold take place.

🌿 implement `.shouldEvalPolicy()` to determine if a policy should evaluate a given resource.
🌿 expose `.policyConfigSchema` as a simple policy configuration schema
🌿 add interface `PolicyConfigSchemaArgs` interface
🐛 add missing comment for `PolicyManagerStats` interface